requirements.txt
httpx
yfinance
FastAPI
get-all-tickers
python-decouple
pandas
pytest